requirements-test.txt
# Testing Dependencies for Mishin Learn Platform
# Install with: pip install -r requirements-test.txt

# Core testing frameworks
pytest==7.4.3
pytest-playwright==0.5.2
playwright==1.49.0

# Pytest plugins
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-base-url==2.0.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-cov==4.1.0    # Code coverage

# HTTP testing
requests==2.31.0
httpx==0.25.2

# Data validation and mocking
pytest-mock==3.12.0
faker==20.1.0
python-dotenv==1.0.0

# Reporting and utilities
allure-pytest==2.13.2
pytest-json-report==1.5.0

# Linting and formatting (optional)
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Type stubs
types-requests==2.31.0.10
